Updates benchmark

+++ benchmarks/benchmark.h
@@ -66,7 +66,7 @@
         uint64_t cycles_start, cycles_final, cycles_diff;             \
         uint64_t min_diff = (uint64_t)-1;                             \
         int wrong_answer = 0;                                         \
-        for (int i = 0; i < repeat; i++) {                            \
+        for (size_t i = 0; i < repeat; i++) {                         \
             pre;                                                      \
             __asm volatile("" ::: /* pretend to clobber */ "memory"); \
             RDTSC_START(cycles_start);                                \
